scrapper_backend/requirements.txt
Django
djangorestframework
supabase
requests
beautifulsoup4
django-cors-headers